+++ src/lib/wallet.ts
@@ -2,49 +2,29 @@ import chalk from 'chalk';
 import qrcode from 'qrcode-terminal';
 import { Wallet, ethers } from 'ethers';
 import { getPrivateKey } from './keys.js';
+import { networks } from './constants.js';
 
-export const networks = [
-  {
-    name: 'mainnet',
-    url: 'https://rpc.ankr.com/eth',
-  },
-  {
-    name: 'goerli',
-    url: 'https://rpc.ankr.com/eth_goerli',
-  },
-  {
-    name: 'polygon',
-    url: 'https://rpc.ankr.com/polygon',
-  },
-  {
-    name: 'mumbai',
-    url: 'https://rpc.ankr.com/polygon_mumbai',
-  },
-] as const;
-
-export const getWallet = async (): Promise<Wallet> => {
+export const getWallet = async (): Promise<Wallet | null> => {
   const privateKey = await getPrivateKey();
+  if (!privateKey) return null;
+
   return new Wallet(privateKey);
 };
 
 export const displayWallet = async (): Promise<void> => {
   const wallet = await getWallet();
-  console.log('Wallet address:', chalk.green(wallet.address));
+  const address = wallet?.address ?? 'No wallet found';
+  console.log('Wallet address:', chalk.green(address));
   console.log();
-  qrcode.generate(wallet.address);
+  qrcode.generate(address);
   console.log();
   for (const network of networks) {
     try {
       const provider = new ethers.JsonRpcProvider(network.url);
-      const balance = ethers.formatEther(
-        await provider.getBalance(wallet.address),
-      );
+      const balance = ethers.formatEther(await provider.getBalance(address));
       console.log('📡 --- ' + network.name + ' ---- 📡 ');
       console.log('balance:', balance);
-      console.log(
-        'nonce: ',
-        await provider.getTransactionCount(wallet.address),
-      );
+      console.log('nonce: ', await provider.getTransactionCount(address));
       console.log();
     } catch (error) {
       console.log(' -- ' + network.name + ' --  -- -- 📡 ');
